Module Irmin_pack

module Pack : sig ... end

module Dict : sig ... end

module Index : sig ... end

module Config : sig ... end

module Inode : sig ... end

module type VERSION = sig ... end

val config : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> ?⁠index_log_size:int -> ?⁠merge_throttle:Config.merge_throttle -> ?⁠freeze_throttle:Config.freeze_throttle -> string -> Irmin.config

Configuration options for stores.

parameter fresh

whether an existing store should be overwritten.

parameter read_only

whether read-only mode is enabled for this store.

parameter lru_size

the maximum number of bindings in the lru cache.

parameter index_log_size

the maximum number of bindings in the index cache.

parameter index_throttle

the strategy to use when the index cache is full and an async Index.merge in already in progress. Block_writes (the default) blocks any new writes until the merge is completed. Overcommit_memory does not block but indefinitely expands the in-memory cache.

exception RO_Not_Allowed

exception Unsupported_version of [ `V1 | `V2 ]

module Make_ext (_ : sig ... end) (Config : Config.S) (Metadata : Irmin.Metadata.S) (Contents : Irmin.Contents.S) (Path : Irmin.Path.S) (Branch : Irmin.Branch.S) (Hash : Irmin.Hash.S) (N : Irmin.Private.Node.S with type metadata = Metadata.t and type hash = Hash.t and type step = Path.step) (CT : Irmin.Private.Commit.S with type hash = Hash.t) : sig ... end

module type MAKER = functor (Config : Config.S) -> functor (M : Irmin.Metadata.S) -> functor (C : Irmin.Contents.S) -> functor (P : Irmin.Path.S) -> functor (B : Irmin.Branch.S) -> functor (H : Irmin.Hash.S) -> sig ... end

module Make : MAKER

module Make_V2 : MAKER

module KV (Config : Config.S) : Irmin.KV_MAKER

module Atomic_write (K : Irmin.Type.S) (V : Irmin.Hash.S) (_ : sig ... end) : sig ... end

module Stats : sig ... end

module Layout : sig ... end

module Checks : sig ... end

module Store : sig ... end

module Private : sig ... end