Module Inode.Make

Parameters

module H : Irmin.Hash.S

module Node : Irmin.Private.Node.S with type hash = H.t

module Inter : Internal with type hash = H.t and type Val.metadata = Node.metadata and type Val.step = Node.step

module Pack : Content_addressable.S with type key = H.t and type value = Inter.Raw.t

Signature

include Irmin.CONTENT_ADDRESSABLE_STORE with type 'a t = 'a Pack.t with type key = H.t with type value = Inter.Val.t

Content-addressable stores

Content-addressable stores are store where it is possible to read and add new values. Keys are derived from the values raw contents and hence are deterministic.

type 'a t = 'a Pack.t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key = H.t

The type for keys.

type value = Inter.Val.t

The type for raw values.

val mem : [> Irmin.Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin.Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin.Perms.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin.Perms.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

module Key : Irmin.Hash.S with type t = key

Signature for digest hashes, inspired by Digestif.

module Val : sig ... end

val decode_bin_length : string -> int -> int

val batch : Irmin.Perms.read t -> ( [ Irmin.Perms.read | Irmin.Perms.write ] t -> 'a Lwt.t ) -> 'a Lwt.t

val close : _ t -> unit Lwt.t