Module 4-CA.Make

Parameters

module V : Pack_value.S with type hash := key

Signature

include Content_addressable.S with type key = key with type value = V.t

include Irmin.CONTENT_ADDRESSABLE_STORE with type key = key with type value = V.t

Content-addressable stores

Content-addressable stores are store where it is possible to read and add new values. Keys are derived from the values raw contents and hence are deterministic.

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key = key

The type for keys.

type value = V.t

The type for raw values.

val mem : [> Irmin.Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin.Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

val add : 'a t -> value -> key Lwt.t

Overwrite add to work with a read-only database handler.

val unsafe_add : 'a t -> key -> value -> unit Lwt.t

Overwrite unsafe_add to work with a read-only database handler.

val unsafe_append : ensure_unique:bool -> overcommit:bool -> 'a t -> key -> value -> unit

val unsafe_mem : 'a t -> key -> bool

val unsafe_find : check_integrity:bool -> 'a t -> key -> value option

val generation : 'a t -> Optint.Int63.t

The number of times that clear has been called on this store.

val clear_keep_generation : 'a t -> unit Lwt.t

val close : _ t -> unit Lwt.t

val batch : Irmin.Perms.read t -> ([ Irmin.Perms.read | Irmin.Perms.write ] t -> 'a Lwt.t) -> 'a Lwt.t

type index = index

val v : ?fresh:bool -> ?readonly:bool -> ?lru_size:int -> index:index -> string -> Irmin.Perms.read t Lwt.t

val sync : ?on_generation_change:(unit -> unit) -> 'a t -> unit

syncs a readonly instance with the files on disk. The same file instance is shared between several pack instances. Therefore only the first pack instance that checks a generation change, can see it. on_generation_change is a callback for all pack instances to react to a generation change.

val flush : ?index:bool -> ?index_merge:bool -> 'a t -> unit

val version : _ t -> Version.t

val offset : 'a t -> Optint.Int63.t

val clear_caches : 'a t -> unit

clear_cache t clears all the in-memory caches of t. Persistent data are not removed.

val integrity_check : offset:Optint.Int63.t -> length:int -> key -> _ t -> (unit, [ `Wrong_hash | `Absent_value ]) Stdlib.result