Module type Inode.S

include Irmin.CONTENT_ADDRESSABLE_STORE

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin.Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin.Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin.Perms.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin.Perms.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

type index

val v : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> index:index -> string -> Irmin.Perms.read t Lwt.t

val batch : Irmin.Perms.read t -> ([ Irmin.Perms.read | Irmin.Perms.write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

module Key : Irmin.Hash.S with type t = key

module Val : sig ... end

val integrity_check : offset:int64 -> length:int -> key -> _ t -> (unit, [ `Wrong_hash | `Absent_value ]) Stdlib.result

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

val sync : ?⁠on_generation_change:(unit -> unit) -> 'a t -> unit

val clear_caches : 'a t -> unit

val decode_bin : dict:(int -> string option) -> hash:(int64 -> key) -> string -> int -> int

val integrity_check_inodes : [ `Read ] t -> key -> (unit, string) Stdlib.result Lwt.t