Module File.Make

Parameters

module V : ELT with type hash := key

Signature

include Irmin.CONTENT_ADDRESSABLE_STORE with type key = key with type value = V.t

type 'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key = key

The type for keys.

type value = V.t

The type for raw values.

val mem : [> `Read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> `Read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> `Write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> `Write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

type index = index

val v : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> index:index -> string -> [ `Read ] t Lwt.t

val batch : [ `Read ] t -> ([ `Read | `Write ] t -> 'a Lwt.t) -> 'a Lwt.t

val unsafe_append : 'a t -> key -> value -> unit

val unsafe_mem : 'a t -> key -> bool

val unsafe_find : 'a t -> key -> value option

val sync : 'a t -> unit

type integrity_error = [

| `Wrong_hash

| `Absent_value

]

val integrity_check : offset:int64 -> length:int -> key -> 'a t -> (unit, integrity_error) Stdlib.result

val close : 'a t -> unit Lwt.t