Module type Pack_value.S

include Irmin.Type.S

type t

val t : t Repr__Type.ty

type hash

type key

val hash : t -> hash

val kind : t -> Kind.t

val length_header : Kind.t -> [ `Varint ] option

Describes the length header formats for the data sections of pack entries.

val weight : t -> Irmin_pack__.Pack_value_intf.weight

weight t is the t's LRU weight.

val encode_bin : dict:(string -> int option) -> offset_of_key:(key -> Optint.Int63.t option) -> hash -> t Irmin.Type.encode_bin

val decode_bin : dict:(int -> string option) -> key_of_offset:(Optint.Int63.t -> key) -> key_of_hash:(hash -> key) -> t Irmin.Type.decode_bin

val decode_bin_length : string -> int -> int

val to_kinded : t -> Irmin_pack__.Pack_value_intf.kinded

to_kinded t returns a kinded version of t.

val of_kinded : Irmin_pack__.Pack_value_intf.kinded -> t

of_kinded k is the inverse of to_kinded t.

It is expected that an implementation only works for k that is returned from to_kinded t and will raise an exception otherwise.