Module Stats.Find

type location =

| Staging	(*
Found in the store's write buffer.
*)

| Lru	(*
Found in the store's LRU of recent find results.
*)

| Pack_direct	(*
Decoded directly from the pack file (via a direct key).
*)

| Pack_indexed	(*
Binding recovered from the pack file after first checking the index for its offset and length (via an indexed key).
*)

| Not_found	(*
Find returned None.
*)

val location_t : location Irmin.Type.t

type t = {

mutable total : int;

mutable from_staging : int;

mutable from_lru : int;

mutable from_pack_direct : int;

mutable from_pack_indexed : int;

}

val t : t Irmin.Type.t

val cache_misses : t -> int