Module Stats.Inode

type field =

| Inode_add

| Inode_remove

| Inode_of_seq

| Inode_of_raw

| Inode_rec_add

| Inode_rec_remove

| Inode_to_binv

| Inode_decode_bin

| Inode_encode_bin

type t = private {

mutable inode_add : int;

mutable inode_remove : int;

mutable inode_of_seq : int;

mutable inode_of_raw : int;

mutable inode_rec_add : int;

mutable inode_rec_remove : int;

mutable inode_to_binv : int;

mutable inode_decode_bin : int;

mutable inode_encode_bin : int;

}

The type for stats for a store S.

inode_add + inode_remove + inode_of_seq + inode_of_raw is the total number of Inode.Val.t built;

inode_rec_add + inode_rec_remove are witnesses of the quantity of work that is done modifying inodes;

inode_to_binv is the number of Inode.Bin.v built;

inode_encode_bin is the number of Bin to Compress conversions;

inode_decode_bin is the number of Compress to Bin conversions;

val t : t Irmin.Type.t

type stat

val export : stat -> t