Module Make.Snapshot

type kinded_hash =

| Contents of hash * metadata

| Node of hash

val kinded_hash_t : kinded_hash Irmin.Type.t

type entry = {

step : string;

hash : kinded_hash;

}

val entry_t : entry Irmin.Type.t

type inode_tree = {

depth : int;

length : int;

pointers : (int * hash) list;

}

val inode_tree_t : inode_tree Irmin.Type.t

type v =

| Inode_tree of inode_tree

| Inode_value of entry list

val v_t : v Irmin.Type.t

type inode = {

v : v;

root : bool;

}

val inode_t : inode Irmin.Type.t

type t =

| Inode of inode

| Blob of Backend.Contents.Val.t

val t : t Irmin.Type.t

val export : ?on_disk:[ `Path of string ] -> repo -> ( t -> unit Lwt.t ) -> root_key:Tree.kinded_key -> int Lwt.t

export ?on_disk repo f ~root_key applies f to all inodes and contents in a rooted tree, with root specified by root_key.

The traversal requires an index to keep track of visited elements.

if on_disk is not specified, the index is in memory.

if on_disk is `Path path, a temporary index is created at path.

The traversal order is stable. In Inode_tree, it is lexicographic on the index function (see Conf.inode_child_order). In Inode_value, it is lexicographic on the steps.

f is called in post-order, that is f is first called on the leaves, and the last call to f is on the root designated by root_key.

The traversal skips objects that are structurally equal to objects that were already traversed. In other words, export internally uses a hash set in order to guarantee that all the objects passed to f don't hash the same way.

Returns the total number of elements visited.

module Import : sig ... end