Module type Irmin_pack.Specifics

type repo

type commit

val integrity_check : ?ppf:Stdlib.Format.formatter -> auto_repair:bool -> repo -> ( [> `Fixed of int | `No_error ], [> `Cannot_fix of string | `Corrupted of int ] ) Stdlib.result

Checks the integrity of the repository. if auto_repair is true, will also try to fix the issues. ppf is a formatter for progressive reporting. `Fixed and `Corrupted report the number of fixed/corrupted entries.

val sync : repo -> unit

sync t syncs a readonly pack with the files on disk. Raises invalid_argument if called by a read-write pack.

val flush : repo -> unit

flush t flush read-write pack on disk. Raises RO_Not_Allowed if called by a readonly instance.