Module type Irmin_pack_intf.Sigs

module Dict = Irmin_pack__.Pack_dict

module Index = Irmin_pack__.Pack_index

module Conf = Irmin_pack.Conf

module Inode = Irmin_pack.Inode

module Pack_value = Irmin_pack.Pack_value

module Pack_store = Irmin_pack.Pack_store

module Version = Irmin_pack.Version

val config : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> ?⁠index_log_size:int -> ?⁠merge_throttle:Conf.merge_throttle -> ?⁠freeze_throttle:Conf.freeze_throttle -> string -> Irmin.config

Configuration options for stores.

parameter fresh

whether an existing store should be overwritten.

parameter read_only

whether read-only mode is enabled for this store.

parameter lru_size

the maximum number of bindings in the lru cache.

parameter index_log_size

the maximum number of bindings in the index cache.

parameter index_throttle

the strategy to use when the index cache is full and an async Index.merge in already in progress. Block_writes (the default) blocks any new writes until the merge is completed. Overcommit_memory does not block but indefinitely expands the in-memory cache.

exception RO_not_allowed

module Make : functor (_ : Version.S) -> Maker

module V1 : Maker

module V2 : Maker

module KV : functor (_ : Version.S) -> functor (_ : Conf.S) -> Irmin.KV_MAKER

module type S = Irmin_pack__.S.S

module type Specifics = Irmin_pack__.S.Specifics

module Make_ext : functor (_ : Version.S) -> functor (_ : Conf.S) -> functor (N : Irmin.Private.Node.Maker) -> functor (CT : Irmin.Private.Commit.Maker) -> Irmin_pack__.S.Maker

module Stats = Irmin_pack.Stats

module Layout = Irmin_pack.Layout

module Checks = Irmin_pack.Checks

val migrate : Irmin.config -> unit

migrate conf upgrades the repository with configuration conf to use the latest storage format.

Note: performing concurrent store operations during the migration, or attempting to use pre-migration instances of the repository after the migration is complete, will result in undefined behaviour.

module Content_addressable = Irmin_pack.Content_addressable

module Atomic_write = Irmin_pack.Atomic_write

module IO = Irmin_pack.IO

module Utils = Irmin_pack.Utils

module type Maker = functor (_ : Conf.S) -> Irmin_pack__.S.Maker