Module type Pack_intf.S

include Irmin.CONTENT_ADDRESSABLE_STORE

type 'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> `Read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> `Read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> `Write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> `Write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

val add : 'a t -> value -> key Lwt.t

Overwrite add to work with a read-only database handler.

val unsafe_add : 'a t -> key -> value -> unit Lwt.t

Overwrite unsafe_add to work with a read-only database handler.

type index

val v : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> index:index -> string -> [ `Read ] t Lwt.t

val batch : [ `Read ] t -> ([ `Read | `Write ] t -> 'a Lwt.t) -> 'a Lwt.t

val unsafe_append : ensure_unique:bool -> overcommit:bool -> 'a t -> key -> value -> unit

val unsafe_mem : 'a t -> key -> bool

val unsafe_find : check_integrity:bool -> 'a t -> key -> value option

val flush : ?⁠index:bool -> ?⁠index_merge:bool -> 'a t -> unit

val sync : ?⁠on_generation_change:(unit -> unit) -> 'a t -> unit

syncs a readonly instance with the files on disk. The same file instance is shared between several pack instances. Therefore only the first pack instance that checks a generation change, can see it. on_generation_change is a callback for all pack instances to react to a generation change.

val version : 'a t -> Irmin_pack__.IO.version

val generation : 'a t -> int64

val offset : 'a t -> int64

val clear : 'a t -> unit Lwt.t

clear t removes all the data from t.

val clear_caches : 'a t -> unit

clear_cache t clears all the in-memory caches of t. Persistent data are not removed.

include Sigs.CHECKABLE with type 'a t := 'a t and type key := key

type 'a t

type key

val integrity_check : offset:int64 -> length:int -> key -> _ t -> (unit, Sigs.integrity_error) Stdlib.result

include Sigs.CLOSEABLE with type 'a t := 'a t

type 'a t

val close : _ t -> unit Lwt.t

val clear_keep_generation : 'a t -> unit Lwt.t