Parameter Closeable.1-CA

include Irmin.CONTENT_ADDRESSABLE_STORE

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin__.Import.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin__.Import.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

include Irmin__.S.Store_properties.CLEARABLE with type 'a t := 'a t

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

val add : 'a t -> value -> key Lwt.t

Overwrite add to work with a read-only database handler.

val unsafe_add : 'a t -> key -> value -> unit Lwt.t

Overwrite unsafe_add to work with a read-only database handler.

val unsafe_append : ensure_unique:bool -> overcommit:bool -> 'a t -> key -> value -> unit

val unsafe_mem : 'a t -> key -> bool

val unsafe_find : check_integrity:bool -> 'a t -> key -> value option

val generation : 'a t -> Irmin_pack__.Import.int63

The number of times that clear has been called on this store.

val clear_keep_generation : 'a t -> unit Lwt.t

val close : _ t -> unit Lwt.t

val batch : Irmin_pack__.Import.read t -> ([ Irmin_pack__.Import.read | Irmin_pack__.Import.write ] t -> 'a Lwt.t) -> 'a Lwt.t