Module 1-Args.Inode

include Irmin.CONTENT_ADDRESSABLE_STORE

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin__.Import.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin__.Import.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

include Irmin__.S.Store_properties.CLEARABLE with type 'a t := 'a t

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

module Key : Irmin.Hash.S with type t = key

module Val : Irmin_pack__.Inode_intf.Value with type t = value and type hash = key

val decode_bin_length : string -> int -> int

val batch : Irmin_pack__.Import.read t -> ([ Irmin_pack__.Import.read | Irmin_pack__.Import.write ] t -> 'a Lwt.t) -> 'a Lwt.t

val close : _ t -> unit Lwt.t