Module type Checks.S

include Irmin_pack.Checks.S

module Stat : sig ... end

Reads basic metrics from an existing store and prints them to stdout.

module Reconstruct_index : sig ... end

Rebuilds an index for an existing pack file

module Integrity_check : sig ... end

Checks the integrity of a store

module Integrity_check_index : sig ... end

Checks the integrity of the index in a store

module Integrity_check_inodes : sig ... end

Checks the integrity of inodes in a store

module Check_self_contained : sig ... end

val cli : unit -> Irmin_pack.Checks.empty

Run a Cmdliner binary containing tools for running offline checks.