Module type S.Content_addressable

include Irmin_pack.Pack_store.S

include Irmin_pack.Content_addressable.S

include Irmin.CONTENT_ADDRESSABLE_STORE

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin__.Import.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin__.Import.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

include Irmin__.S.Store_properties.CLEARABLE with type 'a t := 'a t

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

val add : 'a t -> value -> key Lwt.t

Overwrite add to work with a read-only database handler.

val unsafe_add : 'a t -> key -> value -> unit Lwt.t

Overwrite unsafe_add to work with a read-only database handler.

val unsafe_append : ensure_unique:bool -> overcommit:bool -> 'a t -> key -> value -> unit

val unsafe_mem : 'a t -> key -> bool

val unsafe_find : check_integrity:bool -> 'a t -> key -> value option

val generation : 'a t -> Irmin_pack__.Import.int63

The number of times that clear has been called on this store.

val clear_keep_generation : 'a t -> unit Lwt.t

val close : _ t -> unit Lwt.t

val batch : Irmin_pack__.Import.read t -> ([ Irmin_pack__.Import.read | Irmin_pack__.Import.write ] t -> 'a Lwt.t) -> 'a Lwt.t

type index

val v : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> index:index -> string -> Irmin_pack__.Import.read t Lwt.t

val sync : ?⁠on_generation_change:(unit -> unit) -> 'a t -> unit

syncs a readonly instance with the files on disk. The same file instance is shared between several pack instances. Therefore only the first pack instance that checks a generation change, can see it. on_generation_change is a callback for all pack instances to react to a generation change.

val flush : ?⁠index:bool -> ?⁠index_merge:bool -> 'a t -> unit

val version : _ t -> Irmin_pack.Version.t

val offset : 'a t -> Irmin_pack__.Import.int63

val clear_caches : 'a t -> unit

clear_cache t clears all the in-memory caches of t. Persistent data are not removed.

type 'a t

type key

val integrity_check : offset:Irmin_pack__.Import.int63 -> length:int -> key -> _ t -> (unit, [ `Wrong_hash | `Absent_value ]) Stdlib.result

module U : Irmin_pack.Pack_store.S with type value = value and type index := index

module L : Irmin_pack.Pack_store.S with type index := index

val v : Irmin_pack_layered__.Import.read U.t -> Irmin_pack_layered__.Import.read U.t -> Irmin_pack_layered__.Import.read L.t option -> flip:bool -> freeze_in_progress:(unit -> bool) -> Irmin_pack_layered__.Import.read t

val layer_id : Irmin_pack_layered__.Import.read t -> key -> Irmin_layers.Layer_id.t Lwt.t

type 'a layer_type =

| Upper : Irmin_pack_layered__.Import.read U.t layer_type

| Lower : Irmin_pack_layered__.Import.read L.t layer_type

val copy : ('l layer_type * 'l) -> Irmin_pack_layered__.Import.read t -> string -> key -> unit

val copy_from_lower : Irmin_pack_layered__.Import.read t -> dst:'a U.t -> ?⁠aux:(value -> unit Lwt.t) -> string -> key -> unit Lwt.t

val mem_lower : 'a t -> key -> bool Lwt.t

val mem_next : [> Irmin_pack_layered__.Import.read ] t -> key -> bool Lwt.t

val current_upper : 'a t -> Irmin_pack_layered__.Import.read U.t

val next_upper : 'a t -> Irmin_pack_layered__.Import.read U.t

val lower : 'a t -> Irmin_pack_layered__.Import.read L.t

val clear_previous_upper : ?⁠keep_generation:unit -> 'a t -> unit Lwt.t

val sync : ?⁠on_generation_change:(unit -> unit) -> ?⁠on_generation_change_next_upper:(unit -> unit) -> 'a t -> bool

include Layered_general with type 'a t := 'a t

type 'a t

val close : _ t -> unit Lwt.t

val update_flip : flip:bool -> _ t -> unit

val flip_upper : _ t -> unit

val clear_caches_next_upper : 'a t -> unit

val unsafe_append : ensure_unique:bool -> overcommit:bool -> 'a t -> key -> value -> unit

val flush_next_lower : 'a t -> unit

val integrity_check : offset:Irmin_pack_layered__.Import.int63 -> length:int -> layer:Irmin_layers.Layer_id.t -> key -> _ t -> (unit, Irmin_pack.Checks.integrity_error) Stdlib.result

val consume_newies : 'a t -> key list

val check : 'a t -> ?⁠none:(unit -> unit Lwt.t) -> ?⁠some:(value -> unit Lwt.t) -> key -> unit Lwt.t