Module Irmin_pack_mem

A fully in-memory implementation of the Irmin_pack flavour of Irmin backend, intended for users that must be interoperable with the idiosyncrasies of the persistent implementation.

module Make : functor (_ : Irmin.Private.Node.Maker) -> functor (_ : Irmin.Private.Commit.Maker) -> Irmin_pack.Maker