Parameter Maker.1-_

val entries : int

val stable_hash : int

val contents_length_header : Irmin_pack.Conf.length_header

Describes the length header of the user's contents values when binary-encoded. Supported modes are:

Some `Varint: the length header is a LEB128-encoded integer at the very beginning of the encoded value.

None: there is no length header, and values have unknown size. NOTE: when using irmin-pack in this mode, the selected indexing strategy must index all contents values (as recovering contents values from the store will require referring to the index for their length information).

val inode_child_order : Irmin_pack.Conf.inode_child_order