Module Irmin_pack_unix

The irmin-pack-unix package provides an implementation of Irmin_pack for Unix systems.

irmin-pack-unix provides advanced features such as garbage collection, snapshoting, integrity checks.

Store

module type S = sig ... end

An irmin-pack-unix store. This provides the common Irmin interface with irmin-pack-unix specific extensions.

module Maker (Config : Irmin_pack.Conf.S) : sig ... end

module KV (Config : Irmin_pack.Conf.S) : sig ... end

Key and Values

module Pack_key : sig ... end

module Pack_value : sig ... end

Integrity Checks

module Checks : sig ... end

Offline stats for Irmin stores.

Statistics

module Stats : sig ... end

Internal Functors and Utilities

Following functors and modules are instantiated automatically or used internally when creating a store with Maker or KV.

module Index : sig ... end

module Inode : sig ... end

module Pack_store : sig ... end

module Io_legacy : sig ... end

module Atomic_write : sig ... end

module Dict : sig ... end

module Dispatcher : sig ... end

module Io : sig ... end

module Async : sig ... end

module Errors : sig ... end

module Io_errors : sig ... end

module Control_file : sig ... end

module Append_only_file : sig ... end

module Chunked_suffix : sig ... end

module File_manager : sig ... end

module Mapping_file : sig ... end

module Utils : sig ... end