Module Checks.Make

Parameters

module _ : sig ... end

Signature

module Stat : sig ... end

Reads basic metrics from an existing store and prints them to stdout.

module Reconstruct_index : sig ... end

Rebuilds an index for an existing pack file

module Integrity_check : sig ... end

Checks the integrity of a store

module Integrity_check_index : sig ... end

Checks the integrity of the index in a store

module Integrity_check_inodes : sig ... end

Checks the integrity of inodes in a store

module Stats_commit : sig ... end

Traverses a commit to get stats on its underlying tree.

val cli : ?terms:(unit Cmdliner.Term.t * Cmdliner.Term.info) list -> unit -> Irmin_pack_unix__.Checks_intf.empty

Run a Cmdliner binary containing tools for running offline checks. terms defaults to the set of checks in this module.