Module _.Backend

Backend functions, which might be used by the backends.

module Schema = Schema

A store schema, meant to be provided by the user.

module Hash = Hash

Hashing implementation.

module Contents : Irmin.Contents.Store with type hash = Hash.t and type value = Schema.Contents.t with type key = contents_key

A contents store.

module Node : Irmin.Node.Store with type hash = Hash.t and type Val.contents_key = Contents.key and module Path = Schema.Path and module Metadata = Schema.Metadata with module Path = Path with type key = node_key

A node store.

module Node_portable : sig ... end

A node abstraction that is portable from different repos. Similar to Node.Val.

module Commit : Irmin.Commit.Store with type hash = Hash.t and type Val.node_key = Node.key and module Info = Schema.Info with type key = commit_key

A commit store.

module Commit_portable : sig ... end

A commit abstraction that is portable from different repos. Similar to Commit.Val.

module Branch : Irmin.Branch.Store with type key = Schema.Branch.t and type value = Commit.key

A branch store.

module Slice : sig ... end

A slice abstraction.

module Repo : sig ... end

A repo abstraction.

module Remote : sig ... end

URI-based low-level remote synchronisation.