Module Sparse.Wo

type t

val open_wo : mapping_size:int -> mapping:string -> data:string -> (t, [> open_error ]) Stdlib.result

open_wo ~mapping_size ~mapping ~data returns a write-only instance of the sparse file.

Note: This is unsafe and is only used by the GC to mark the parent commits as dangling. One must ensure that no read-only instance is opened at the same time, as otherwise the writes would be observable by it.

val write_exn : t -> off:Optint.Int63.t -> len:int -> string -> unit

write_exn t ~off ~len str writes the first len bytes of str to t at the virtual offset off.

val fsync : t -> (unit, [> Io.write_error ]) Stdlib.result

fsync t persists to the file system the effects of previous writes.

val close : t -> (unit, [> Io.close_error ]) Stdlib.result

Close the underlying files.

val create_from_data : mapping:string -> dead_header_size:int -> size:Optint.Int63.t -> data:string -> (Optint.Int63.t, [> Io.create_error | Io.write_error | Io.close_error ]) Stdlib.result

create_from_data ~mapping ~dead_header_size ~size ~data initializes a new sparse file on disk from the existing file data, by creating the corresponding mapping file. The first dead_header_size bytes are ignored and the remaining size bytes of data are made available.

On success, returns the size of the mapping file to be stored in the control file for consistency checking on open.