Module Irmin_pack_unix.Chunked_suffix

module type S = sig ... end

Abstraction for a chunked suffix. It is functionally equivalent to Append_only_file but with a chunked implementation that is parameterized by

module Make (Io : Io.S) (Errs : Io_errors.S with module Io = Io) : S with module Io = Io and module Errs = Errs

Abstraction for a chunked suffix. It is functionally equivalent to Append_only_file but with a chunked implementation that is parameterized by