Module Upper.V5

type gced = {

suffix_start_offset : Optint.Int63.t;

generation : int;

latest_gc_target_offset : Optint.Int63.t;

suffix_dead_bytes : Optint.Int63.t;

mapping_end_poff : Optint.Int63.t option;

}

val gced_t : gced Irmin.Type.t

type status =

| From_v1_v2_post_upgrade of V3.from_v1_v2_post_upgrade

| No_gc_yet

| Used_non_minimal_indexing_strategy

| Gced of gced

| T1

| T2

| T3

| T4

| T5

| T6

| T7

| T8

| T9

| T10

| T11

| T12

| T13

| T14

| T15

val status_t : status Irmin.Type.t

type t = {

dict_end_poff : Optint.Int63.t;

appendable_chunk_poff : Optint.Int63.t;

upgraded_from : int option;

checksum : Optint.Int63.t;

chunk_start_idx : int;

chunk_num : int;

volume_num : int;

status : status;

}

The same as V4.t, with the following modifications:

New fields

volume_num stores the number of volumes in the lower layer.

mapping_end_poff stores the mapping file size (optional if missing or unknown after a migration from V4).

Changed fields

Replaced upgraded_from_v3_to_v4 with generic field upgraded_from to track version upgrades. Note, it is an int option since Version.t option has a bug somewhere in repr that needs further investigation.

val t : t Irmin.Type.t