Module Volume.V5

type t = {

start_offset : Optint.Int63.t;

end_offset : Optint.Int63.t;

mapping_end_poff : Optint.Int63.t;

checksum : Optint.Int63.t;

}

The payload for a control file of a volume.

Fields

start_offset is the global offset for the start of the volume's data. Used for routing reads.

end_offset is the global offset for the end of the volume's data. Used for routing reads.

mapping_end_poff is the end offset for the mapping file. Used when writing.

val t : t Irmin.Type.t