Module type Dict.S

module Fm : File_manager.S

Abstraction that governs the lifetime of the various files that are part of a pack store (except the branch store).

type t

val find : t -> int -> string option

val index : t -> string -> int option

val v : Fm.t -> ( t, [> Fm.Io.read_error ] ) Stdlib.result

val close : t -> unit