Module Irmin_pack_unix.File_manager

module type S = sig ... end

Abstraction that governs the lifetime of the various files that are part of a pack store (except the branch store).

module Make (Control : Control_file.S with module Io = Io.Unix) (Dict : Append_only_file.S with module Io = Control.Io) (Suffix : Append_only_file.S with module Io = Control.Io) (Index : sig ... end) (Errs : Io_errors.S with module Io = Control.Io) : S with module Io = Control.Io and module Control = Control and module Dict = Dict and module Suffix = Suffix and module Index = Index

Abstraction that governs the lifetime of the various files that are part of a pack store (except the branch store).