Module S.Control

module Io = Io

type t

val create_rw : path:string -> tmp_path:string option -> overwrite:bool -> Irmin_pack_unix__.Control_file_intf.Payload.Upper.Latest.t -> (t, [> Io.create_error | Io.write_error ]) Stdlib.result

Create a rw instance of t by creating a control file.

type open_error := [

| `Corrupted_control_file of string

| `Io_misc of Io.misc_error

| `No_such_file_or_directory of string

| `Not_a_file

| `Closed

| `Unknown_major_pack_version of string

]

val open_ : path:string -> tmp_path:string option -> readonly:bool -> (t, [> open_error ]) Stdlib.result

Create a rw instance of t by reading an existing file at path. tmp_path will be used by RW instances when updating it's content, it is not required for RO instances or RW instances which will never be updated.

val close : t -> (unit, [> Io.close_error ]) Stdlib.result

val read_payload : path:string -> (Irmin_pack_unix__.Control_file_intf.Payload.Upper.Latest.t, [> open_error | Io.close_error ]) Stdlib.result

read_payload ~path reads the payload at path. It is a convenient way to read the payload without needing to call open_, payload, close.

val read_raw_payload : path:string -> (Irmin_pack_unix__.Control_file_intf.Payload.Upper.raw_payload, [> open_error | Io.close_error ]) Stdlib.result

val payload : t -> Irmin_pack_unix__.Control_file_intf.Payload.Upper.Latest.t

payload t is the payload in t.

That function doesn't perform IO.

RW mode

payload t is the payload, as it was written to the file system.

RO mode

payload t is the payload, as it was seen during open_ or during the most recent reload.

type reload_error := [

| `Rw_not_allowed

| open_error

| Io.close_error

]

val reload : t -> (unit, [> reload_error ]) Stdlib.result

RW mode

Always returns an error.

RO mode

Reread the file on disk.

If the file changed since the last read, the payload in t is updated to match the content of the file.

type move_error := [

| `Sys_error of string

]

type set_error := [

| `No_tmp_path_provided

| Io.create_error

| Io.write_error

| move_error

| Io.close_error

]

val set_payload : t -> Irmin_pack_unix__.Control_file_intf.Payload.Upper.Latest.t -> (unit, [> set_error ]) Stdlib.result

RW mode

Write a new payload on disk.

RO mode

Always returns an error.

val readonly : t -> bool

val fsync : t -> (unit, [> Io.write_error ]) Stdlib.result

RW mode

Tell the OS to fush its internal buffers.

RO mode

Always returns Error `Ro_not_allowed.