Module Irmin_pack_unix.Index

module type S = sig ... end

An abstraction on top of the index library that exposes an API that better fits the irmin-pack use case.

module Make (K : Irmin.Hash.S) : S with type key = K.t and module Io = Io.Unix

An abstraction on top of the index library that exposes an API that better fits the irmin-pack use case.