Module Pack.Entry_prefix

type t = {

hash : hash;

kind : Irmin_pack.Pack_value.Kind.t;

size_of_value_and_length_header : int option;
(*
Remaining bytes in the entry after reading the hash and the kind (i.e. the length of the length header + the value of the length header), if the entry has a length header (otherwise None).

NOTE: the length stored in the index and in direct pack keys is the total_entry_length (including the hash and the kind).
*)

}

val total_entry_length : t -> int option