Module Make.Internal

Unstable internal API agnostic about the underlying storage. Use it only to implement or test inodes.

module Io = Io.Unix

module Errs : Io_errors.S with module Io = Io

module Index : sig ... end

An abstraction on top of the index library that exposes an API that better fits the irmin-pack use case.

module File_manager : File_manager.S with module Io = Io and module Errs = Errs and module Index = Index

Abstraction that governs the lifetime of the various files that are part of a pack store (except the branch store).

val file_manager : repo -> File_manager.t

module Dict : Dict.S

val dict : repo -> Dict.t

module XKey : Pack_key.S with type hash = Schema.Hash.t

val suffix_commit_mem : repo -> XKey.t -> bool

val suffix_node_mem : repo -> XKey.t -> bool

val suffix_contents_mem : repo -> XKey.t -> bool

val kill_gc : repo -> bool