Module Lower.Make_volume

Parameters

module Io : Io.S

module Errs : Io_errors.S with module Io = Io

Signature

module Io = Io

module Errs = Errs

module Sparse : Sparse_file.S

type t

type open_error = [

| Io.open_error

| `Closed

| `Double_close

| `Corrupted_control_file of string

| `Unknown_major_pack_version of string

]

val v : string -> (t, [> open_error ]) Stdlib.result

v path loads the volume at path in read-only.

val path : t -> string

path t is the directory that contains the volume.

val is_empty : t -> bool

is_empty t returns whether t is empty or not.

val control : t -> Control_file.Payload.Volume.Latest.t option

control t returns the control file payload for the volume.

val identifier : t -> string

identifier t is a unique idendifier for the volume.