Module Irmin_pack_unix.Lru

type t

An LRU that support memory-bound capacity via configuration key lru_max_memory. Falls back to entry-based capacity via lru_size configuration key, if max memory is not configured.

type key = Optint.Int63.t

type value = Irmin_pack.Pack_value.kinded

val create : Irmin.Backend.Conf.t -> t

val add : t -> Optint.Int63.t -> Irmin_pack.Pack_value.weight -> value -> unit

add t key weight value maps value with weight to key in t.

Note: Irmin_pack.Pack_value.weight.Immediate weights will be checked to see if they exceed an entry weight limit. The current hard-coded entry weight limit is 20kB.

val find : t -> key -> value

val mem : t -> key -> bool

val clear : t -> unit

val iter : t -> (key -> value -> unit) -> unit