Module type Pack_store.S

include Irmin_pack.Indexable.S

include Irmin.Indexable.S

Read-only stores

Read-only stores are store where it is only possible to read existing values.

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin.Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin.Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

type hash

The type of hashes of value.

val index : [> Irmin.Perms.read ] t -> hash -> key option Lwt.t

Indexing maps the hash of a value to a corresponding key of that value in the store. For stores that are addressed by hashes directly, this is typically fun _t h -> Lwt.return (Key.of_hash h); for stores with more complex addressing schemes, index may attempt a lookup operation in the store.

In general, indexing is best-effort and reveals no information about the membership of the value in the store. In particular:

index t hash = Some key doesn't guarantee mem t key: the value with hash hash may still be absent from the store;

index t hash = None doesn't guarantee that there is no key such that mem t key and Key.to_hash key = hash: the value may still be present in the store under a key that is not indexed.

val batch : Irmin.Perms.read t -> ([ Irmin.Perms.read | Irmin.Perms.write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

module Key : Irmin.Key.S with type t = key and type hash = hash

val add : 'a t -> value -> key Lwt.t

Overwrite add to work with a read-only database handler.

val unsafe_add : 'a t -> hash -> value -> key Lwt.t

Overwrite unsafe_add to work with a read-only database handler.

val index_direct : _ t -> hash -> key option

val unsafe_append : ensure_unique:bool -> overcommit:bool -> 'a t -> hash -> value -> key

val unsafe_mem : 'a t -> key -> bool

val unsafe_find : check_integrity:bool -> 'a t -> key -> value option

type file_manager

type dict

type dispatcher

val v : config:Irmin.Backend.Conf.t -> fm:file_manager -> dict:dict -> dispatcher:dispatcher -> Irmin.Perms.read t

val cast : Irmin.Perms.read t -> Irmin.Perms.read_write t

val integrity_check : offset:Optint.Int63.t -> length:int -> hash -> _ t -> (unit, [ `Wrong_hash | `Absent_value ]) Stdlib.result

module Entry_prefix : sig ... end

val read_and_decode_entry_prefix : off:Optint.Int63.t -> dispatcher -> Entry_prefix.t

Read the entry prefix at offset off.

val index_direct_with_kind : 'a t -> hash -> (key * Irmin_pack.Pack_value.Kind.t) option

Returns the key and the kind of an object indexed by hash.

val purge_lru : 'a t -> unit

val key_of_offset : 'a t -> Optint.Int63.t -> key

Returns the key associated with the offset.

val unsafe_find_no_prefetch : 'a t -> key -> value option

Similar to unsafe_find, returns the value associated with the key in the store but without prefetching the hash and length of the children (or doing any integrity check). As a result, the produced children keys only contain their offset and are not usable without calling key_of_offset first. This function only exists to optimize the GC reachability traversal.