Module Irmin_pack_unix.Pack_value

include module type of struct include Irmin_pack.Pack_value end

module Kind = Irmin_pack.Pack_value.Kind

type nonrec weight =

| Immediate of int

| Deferred of unit -> int

type nonrec kinded = ..

kinded is an extenisble variant that each S extends so that it can define S.to_kinded and S.of_kinded. Its purpose is to allow containers, such as Irmin_pack_unix.Lru, to store and return all types of S and thus be usable by modules defined over S, such as Irmin_pack_unix.Pack_store.

module type S = Irmin_pack.Pack_value.S

module type Config = Irmin_pack.Pack_value.Config

module Of_contents = Irmin_pack.Pack_value.Of_contents

module Of_commit = Irmin_pack.Pack_value.Of_commit

module type Persistent = sig ... end