Parameter Of_contents.4-Contents

Signature for store contents

type t

The type for user-defined contents.

val t : t Irmin.Type.t

val merge : t option Irmin.Merge.t

Merge function. Evaluates to `Conflict msg if the values cannot be merged properly. The arguments of the merge function can take None to mean that the key does not exists for either the least-common ancestor or one of the two merging points. The merge function returns None when the key's value should be deleted.