Module Stats.Pack_store

type field =

| Appended_hashes

| Appended_offsets

| Staging	(*
Found in the store's write buffer.
*)

| Lru	(*
Found in the store's LRU of recent find results.
*)

| Pack_direct	(*
Decoded directly from the pack file (via a direct key).
*)

| Pack_indexed	(*
Binding recovered from the pack file after first checking the index for its offset and length (via an indexed key).
*)

| Not_found	(*
Find returned None.
*)

val field_t : field Irmin.Type.t

type t = {

mutable appended_hashes : int;

mutable appended_offsets : int;

mutable total : int;

mutable from_staging : int;

mutable from_lru : int;

mutable from_pack_direct : int;

mutable from_pack_indexed : int;

}

val t : t Irmin.Type.t

type stat

val cache_misses : t -> int

val export : stat -> t