Module Irmin_pack

module Pack : sig ... end

module Dict = Irmin_pack__.Pack_dict

module Index = Irmin_pack__.Pack_index

module Config : sig ... end

val config : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> ?⁠index_log_size:int -> ?⁠merge_throttle:Config.merge_throttle -> ?⁠freeze_throttle:Config.freeze_throttle -> string -> Irmin.config

Configuration options for stores.

parameter fresh

whether an existing store should be overwritten.

parameter read_only

whether read-only mode is enabled for this store.

parameter lru_size

the maximum number of bindings in the lru cache.

parameter index_log_size

the maximum number of bindings in the index cache.

parameter index_throttle

the strategy to use when the index cache is full and an async Index.merge in already in progress. Block_writes (the default) blocks any new writes until the merge is completed. Overcommit_memory does not block but indefinitely expands the in-memory cache.

exception RO_Not_Allowed

exception Unsupported_version of Irmin_pack__.IO.version

module Make_ext : functor (_ : Irmin_pack__.IO.VERSION) -> functor (Config : Config.S) -> functor (Metadata : Irmin.Metadata.S) -> functor (Contents : Irmin.Contents.S) -> functor (Path : Irmin.Path.S) -> functor (Branch : Irmin.Branch.S) -> functor (Hash : Irmin.Hash.S) -> functor (N : Irmin.Private.Node.S with type metadata = Metadata.t and type hash = Hash.t and type step = Path.step) -> functor (CT : Irmin.Private.Commit.S with type hash = Hash.t) -> sig ... end

module type MAKER = functor (Config : Config.S) -> functor (M : Irmin.Metadata.S) -> functor (C : Irmin.Contents.S) -> functor (P : Irmin.Path.S) -> functor (B : Irmin.Branch.S) -> functor (H : Irmin.Hash.S) -> sig ... end

module Make : MAKER

module Make_V2 : MAKER

module KV : functor (Config : Config.S) -> Irmin.KV_MAKER

module Atomic_write : functor (K : Irmin.Type.S) -> functor (V : Irmin.Hash.S) -> functor (_ : Irmin_pack__.IO.VERSION) -> sig ... end

module Stats : sig ... end

module Layout : sig ... end

Defines the location of the IO instances within the main irmin-pack store directory.

module Checks : sig ... end

Offline stats for Irmin stores.

module Store : sig ... end

module Private : sig ... end