Module Irmin_pack

module Conf : sig ... end

module Inode : sig ... end

module Pack_key : sig ... end

module Pack_value : sig ... end

This module defines abstractions over entries in the pack file, which are encoded as follows:

module Pack_store : sig ... end

val config : ?fresh:bool -> ?readonly:bool -> ?lru_size:int -> ?index_log_size:int -> ?merge_throttle:Conf.merge_throttle -> ?freeze_throttle:Conf.freeze_throttle -> ?indexing_strategy:Pack_store.Indexing_strategy.t -> string -> Irmin.config

Configuration options for stores.

parameter fresh
whether an existing store should be overwritten.

parameter read_only
whether read-only mode is enabled for this store.

parameter lru_size
the maximum number of bindings in the lru cache.

parameter index_log_size
the maximum number of bindings in the index cache.

parameter index_throttle
the strategy to use when the index cache is full and an async Index.merge in already in progress. Block_writes (the default) blocks any new writes until the merge is completed. Overcommit_memory does not block but indefinitely expands the in-memory cache.

parameter indexing_strategy
The indexing strategy of the backend store. Defaults to Pack_store.Indexing_strategy.default.

exception RO_not_allowed

module type S = sig ... end

module type Specifics = sig ... end

module type Maker = sig ... end

module type Maker_persistent = sig ... end

module Stats : sig ... end

module Layout : sig ... end

Defines the location of the IO instances within the main irmin-pack store directory.

module Indexable : sig ... end

module Atomic_write : sig ... end

module IO : sig ... end

module Version : sig ... end

Management of disk-format versions.