Module Irmin_pack

module Dict : sig ... end

module Index : sig ... end

module Conf : sig ... end

module Inode : sig ... end

module Pack_value : sig ... end

module Pack_store : sig ... end

module Version : sig ... end

val config : ?fresh:bool -> ?readonly:bool -> ?lru_size:int -> ?index_log_size:int -> ?merge_throttle:Conf.merge_throttle -> ?freeze_throttle:Conf.freeze_throttle -> string -> Irmin.config

Configuration options for stores.

parameter fresh
whether an existing store should be overwritten.

parameter read_only
whether read-only mode is enabled for this store.

parameter lru_size
the maximum number of bindings in the lru cache.

parameter index_log_size
the maximum number of bindings in the index cache.

parameter index_throttle
the strategy to use when the index cache is full and an async Index.merge in already in progress. Block_writes (the default) blocks any new writes until the merge is completed. Overcommit_memory does not block but indefinitely expands the in-memory cache.

exception RO_not_allowed

module Make (_ : Version.S) : sig ... end

module V1 : sig ... end

module V2 : sig ... end

module KV (_ : Version.S) (_ : Conf.S) : Irmin.KV_MAKER

module type S = sig ... end

module type Specifics = sig ... end

module Make_ext (_ : Version.S) (_ : Conf.S) (N : Irmin.Private.Node.Maker) (CT : Irmin.Private.Commit.Maker) : sig ... end

module Stats : sig ... end

module Layout : sig ... end

module Checks : sig ... end

val migrate : Irmin.config -> unit

migrate conf upgrades the repository with configuration conf to use the latest storage format.

Note: performing concurrent store operations during the migration, or attempting to use pre-migration instances of the repository after the migration is complete, will result in undefined behaviour.

module Content_addressable : sig ... end

module Atomic_write : sig ... end

module IO : sig ... end

module Utils : sig ... end

module type Maker = functor (_ : Conf.S) -> sig ... end