Module Inode.Make

Parameters

module H : Irmin.Hash.S

module Key : Irmin.Key.S with type hash = H.t

module Node : Irmin.Node.Generic_key.S with type hash = H.t and type contents_key = Key.t and type node_key = Key.t

module Inter : Internal with type hash = H.t and type key = Key.t and type Snapshot.metadata = Node.metadata and type Val.step = Node.step

module Pack : Indexable.S with type key = Key.t and type hash = H.t and type value = Inter.Raw.t

Signature

include Irmin.Indexable.S with type 'a t = 'a Pack.t with type key = Key.t with type hash = H.t with type value = Inter.Val.t

Read-only stores

Read-only stores are store where it is only possible to read existing values.

type 'a t = 'a Pack.t

The type for stores. The 'a phantom type carries information about the store mutability.

type key = Key.t

The type for keys.

type value = Inter.Val.t

The type for raw values.

val mem : [> Irmin.Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin.Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

type hash = H.t

The type of hashes of value.

val add : [> Irmin.Perms.write ] t -> value -> key Lwt.t

Write the contents of a value to the store, and obtain its key.

val unsafe_add : [> Irmin.Perms.write ] t -> hash -> value -> key Lwt.t

Same as add but allows specifying the value's hash directly. The backend might choose to discard that hash and/or can be corrupt if the hash is not consistent.

val index : [> Irmin.Perms.read ] t -> hash -> key option Lwt.t

Indexing maps the hash of a value to a corresponding key of that value in the store. For stores that are addressed by hashes directly, this is typically fun _t h -> Lwt.return (Key.of_hash h); for stores with more complex addressing schemes, index may attempt a lookup operation in the store.

In general, indexing is best-effort and reveals no information about the membership of the value in the store. In particular:

index t hash = Some key doesn't guarantee mem t key: the value with hash hash may still be absent from the store;

index t hash = None doesn't guarantee that there is no key such that mem t key and Key.to_hash key = hash: the value may still be present in the store under a key that is not indexed.

val batch : Irmin.Perms.read t -> ( [ Irmin.Perms.read | Irmin.Perms.write ] t -> 'a Lwt.t ) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

module Key : Irmin.Key.S with type t = key and type hash = hash

module Hash : Irmin.Hash.S with type t = hash

Signature for digest hashes, inspired by Digestif.

module Val : sig ... end

val decode_bin_length : string -> int -> int

val integrity_check_inodes : [ `Read ] t -> key -> ( unit, string ) Stdlib.result Lwt.t

val save : ?allow_non_root:bool -> 'a t -> value -> key