Module Irmin_pack.Pack_value

This module defines abstractions over entries in the pack file, which are encoded as follows:

      ┌────────┬────────┬──────────────┬─────────┐
      │  Hash  │  Kind  │  Len(Value)? │  Value  │
      └────────┴────────┴──────────────┴─────────┘
      ┆<┄ H ┄┄>┆<┄ K ┄┄>┆<┄┄┄┄ L? ┄┄┄┄>┆<┄┄ V ┄┄>┆
      ┆<┄┄┄┄┄┄┄┄┄┄┄ entry length, E ┄┄┄┄┄┄┄┄┄┄┄┄>┆

The fields are as follows:

Hash: the (fixed-length) hash of the data stored in this entry;

Kind: the kind of data being stored (contents, nodes, commits etc.), encoded as a single "magic" character;

Len(Value): an optional length header for the Value section of the entry (not including the length of the length header itself), encoded using a variable-length encoding (LEB128). The presence of a length header is determined by the Kind character.

Value: the data itself.

The length of the overall pack entry, as referenced in the Pack_index or in a direct Pack_key.t, is equal to E = H + K + L + V.

module Kind : sig ... end

module type S = Irmin_pack__.Pack_value_intf.S with type kind := Kind.t

module type Persistent = Irmin_pack__.Pack_value_intf.Persistent with type kind := Kind.t

module type Config = Irmin_pack__.Pack_value_intf.Config

module Of_contents : functor (_ : Config) -> functor (Hash : Irmin.Hash.S) -> functor (Key : Irmin_pack__.Pack_value_intf.T) -> functor (Contents : Irmin.Contents.S) -> S with type Of_contents.t = Contents.t and type hash = Hash.t and type key = Key.t

module Of_commit : functor (Hash : Irmin.Hash.S) -> functor (Key : Irmin.Key.S with type hash = Hash.t) -> functor (Commit : Irmin.Commit.Generic_key.S with type node_key = Key.t and type commit_key = Key.t) -> S with type Of_commit.t = Commit.t and type hash = Hash.t and type key = Key.t