Module type Irmin_pack_intf.Sigs

module Dict = Irmin_pack__.Pack_dict

module Index = Irmin_pack__.Pack_index

module Conf = Irmin_pack.Conf

module Inode = Irmin_pack.Inode

module Pack_key = Irmin_pack.Pack_key

module Pack_value = Irmin_pack.Pack_value

module Pack_store = Irmin_pack.Pack_store

val config : ?⁠fresh:bool -> ?⁠readonly:bool -> ?⁠lru_size:int -> ?⁠index_log_size:int -> ?⁠merge_throttle:Conf.merge_throttle -> ?⁠freeze_throttle:Conf.freeze_throttle -> ?⁠indexing_strategy:Pack_store.Indexing_strategy.t -> string -> Irmin.config

Configuration options for stores.

parameter fresh

whether an existing store should be overwritten.

parameter read_only

whether read-only mode is enabled for this store.

parameter lru_size

the maximum number of bindings in the lru cache.

parameter index_log_size

the maximum number of bindings in the index cache.

parameter index_throttle

the strategy to use when the index cache is full and an async Index.merge in already in progress. Block_writes (the default) blocks any new writes until the merge is completed. Overcommit_memory does not block but indefinitely expands the in-memory cache.

parameter indexing_strategy

The indexing strategy of the backend store. Defaults to Pack_store.Indexing_strategy.default.

exception RO_not_allowed

module KV : functor (_ : Conf.S) -> Irmin.Generic_key.KV_maker with type metadata = unit and type ('h, 'v) contents_key = 'h Pack_key.t and type 'h node_key = 'h Pack_key.t and type 'h commit_key = 'h Pack_key.t

module type S = S

module type Specifics = Specifics

module type Maker = Maker

module type Maker_persistent = Maker_persistent

module Maker : Maker_persistent

module Stats = Irmin_pack.Stats

module Layout = Irmin_pack.Layout

module Checks = Irmin_pack.Checks

module Indexable = Irmin_pack.Indexable

module Atomic_write = Irmin_pack.Atomic_write

module IO = Irmin_pack.IO

module Version = Irmin_pack.Version