Module Irmin_pack_layered

val config : ?conf:Irmin.config -> ?lower_root:string -> ?upper_root1:string -> ?upper_root0:string -> ?copy_in_upper:bool -> ?with_lower:bool -> ?blocking_copy_size:int -> unit -> Irmin.config

Configuration options for layered stores.

parameter conf
is an irmin-pack configuration.

parameter lower_root
is the root of the lower store, "lower" is the default.

parameter upper_root1
is the root of one of the upper stores, "upper1" is the default.

parameter upper_root0
is the root of one of the upper stores, "upper0" is the default.

parameter copy_in_upper
if true then at the end of a freee the max commits are copied back in upper. This option can be overriden when calling a freeze with the copy_in_upper argument set. By default it is set to false.

parameter with_lower
if true (the default) use a lower layer during freezes.

parameter blocking_copy_size
specifies the maximum size (in bytes) that can be copied in the blocking portion of the freeze.

module type S = sig ... end

module Make_ext (Config : Irmin_pack.Config.S) (Metadata : Irmin.Metadata.S) (Contents : Irmin.Contents.S) (Path : Irmin.Path.S) (Branch : Irmin.Branch.S) (Hash : Irmin.Hash.S) (N : Irmin.Private.Node.S with type metadata = Metadata.t and type hash = Hash.t and type step = Path.step) (CT : Irmin.Private.Commit.S with type hash = Hash.t) : S with type key = Path.t and type contents = Contents.t and type branch = Branch.t and type hash = Hash.t and type step = Path.step and type metadata = Metadata.t and type Key.step = Path.step

module Make (Config : Irmin_pack.Config.S) (M : Irmin.Metadata.S) (C : Irmin.Contents.S) (P : Irmin.Path.S) (B : Irmin.Branch.S) (H : Irmin.Hash.S) : S with type key = P.t and type step = P.step and type metadata = M.t and type contents = C.t and type branch = B.t and type hash = H.t

module Checks : sig ... end