Module Irmin_pack_layered

val config : ?⁠conf:Irmin.config -> ?⁠lower_root:string -> ?⁠upper_root1:string -> ?⁠upper_root0:string -> ?⁠with_lower:bool -> ?⁠blocking_copy_size:int -> unit -> Irmin.config

Configuration options for layered stores.

parameter conf

is an irmin-pack configuration.

parameter lower_root

is the root of the lower store, "lower" is the default.

parameter upper_root1

is the root of one of the upper stores, "upper1" is the default.

parameter upper_root0

is the root of one of the upper stores, "upper0" is the default.

parameter with_lower

if true (the default) use a lower layer during freezes.

parameter blocking_copy_size

specifies the maximum size (in bytes) that can be copied in the blocking portion of the freeze.

module type S = sig ... end

module type Maker = Irmin_pack_layered__.S.Maker

module Maker : functor (_ : Irmin_pack.Conf.S) -> Maker

module Maker_ext : functor (_ : Irmin_pack.Conf.S) -> functor (_ : Irmin.Private.Node.Maker) -> functor (_ : Irmin.Private.Commit.Maker) -> Maker

module Checks : sig ... end