Module Checks.Layout

include module type of struct include Irmin_pack.Layout end

Defines the location of the IO instances within the main irmin-pack store directory.

type path := root:string -> string

val pack : path

val branch : path

val dict : path

val stores : root:string -> string list

val flip : root:string -> string

val lower : root:string -> string

Only works for layered stores that use the default names for layers.

val upper0 : root:string -> string

Only works for layered stores that use the default names for layers.

val upper1 : root:string -> string

Only works for layered stores that use the default names for layers.

val toplevel : string -> string list