Module S.Stat

Reads basic metrics from an existing store and prints them to stdout.

val run : root:string -> unit Lwt.t

val term_internal : (unit -> unit) Cmdliner.Term.t

A pre-packaged Cmdliner term for executing run.

val term : unit Cmdliner.Term.t * Cmdliner.Term.info

term is term_internal plus documentation and logs initialisation

Internal implementation utilities exposed for use in other integrity checks.

type size =

| Bytes of int

val size_t : size Irmin.Type.t

type objects = {

nb_commits : int;

nb_nodes : int;

nb_contents : int;

}

val objects_t : objects Irmin.Type.t

val traverse_index : root:string -> int -> objects