Module Irmin_pack_unix.Errors

val finalise : ( 'a -> 'b ) -> ( unit -> 'c ) -> 'd

Finaliser for a function that returns a result and doesn't raise exceptions.

If the finaliser fails, it is recommended to log the error.

val finalise_exn : ( 'a option -> 'b ) -> ( unit -> 'c ) -> 'd

Finaliser for a function that might raise exceptions.

type base_error = [

| `Double_close

| `File_exists of string

| `Invalid_parent_directory

| `No_such_file_or_directory

| `Not_a_file

| `Read_out_of_bounds

| `Invalid_argument

| `Decoding_error

| `Not_a_directory of string

| `Index_failure of string

| `Invalid_layout

| `Corrupted_legacy_file

| `Corrupted_mapping_file of string

| `Pending_flush

| `Rw_not_allowed

| `Migration_needed

| `Corrupted_control_file

| `Sys_error of string

| `V3_store_from_the_future

| `Gc_forbidden_during_batch

| `Unknown_major_pack_version of string

| `Only_minimal_indexing_strategy_allowed

| `Commit_key_is_dangling of string

| `Dangling_key of string

| `Gc_disallowed

| `Node_or_contents_key_is_indexed of string

| `Commit_parent_key_is_indexed of string

| `Gc_process_error of string

| `Corrupted_gc_result_file of string

| `Gc_process_died_without_result_file of string

| `Gc_forbidden_on_32bit_platforms

| `Invalid_prefix_read of string

| `Invalid_mapping_read of string

| `Invalid_read_of_gced_object of string

| `Inconsistent_store

]

base_error is the type of most errors that can occur in a result, except for errors that have associated exceptions (see below) and backend-specific errors (see Io_errors).

val base_error_t : [ `Commit_key_is_dangling of string | `Commit_parent_key_is_indexed of string | `Corrupted_control_file | `Corrupted_gc_result_file of string | `Corrupted_legacy_file | `Corrupted_mapping_file of string | `Dangling_key of string | `Decoding_error | `Double_close | `File_exists of string | `Gc_disallowed | `Gc_forbidden_during_batch | `Gc_forbidden_on_32bit_platforms | `Gc_process_died_without_result_file of string | `Gc_process_error of string | `Inconsistent_store | `Index_failure of string | `Invalid_argument | `Invalid_layout | `Invalid_mapping_read of string | `Invalid_parent_directory | `Invalid_prefix_read of string | `Invalid_read_of_gced_object of string | `Migration_needed | `No_such_file_or_directory | `Node_or_contents_key_is_indexed of string | `Not_a_directory of string | `Not_a_file | `Only_minimal_indexing_strategy_allowed | `Pending_flush | `Read_out_of_bounds | `Rw_not_allowed | `Sys_error of string | `Unknown_major_pack_version of string | `V3_store_from_the_future ] Irmin.Type.t

val pp_base_error : [ `Commit_key_is_dangling of string | `Commit_parent_key_is_indexed of string | `Corrupted_control_file | `Corrupted_gc_result_file of string | `Corrupted_legacy_file | `Corrupted_mapping_file of string | `Dangling_key of string | `Decoding_error | `Double_close | `File_exists of string | `Gc_disallowed | `Gc_forbidden_during_batch | `Gc_forbidden_on_32bit_platforms | `Gc_process_died_without_result_file of string | `Gc_process_error of string | `Inconsistent_store | `Index_failure of string | `Invalid_argument | `Invalid_layout | `Invalid_mapping_read of string | `Invalid_parent_directory | `Invalid_prefix_read of string | `Invalid_read_of_gced_object of string | `Migration_needed | `No_such_file_or_directory | `Node_or_contents_key_is_indexed of string | `Not_a_directory of string | `Not_a_file | `Only_minimal_indexing_strategy_allowed | `Pending_flush | `Read_out_of_bounds | `Rw_not_allowed | `Sys_error of string | `Unknown_major_pack_version of string | `V3_store_from_the_future ] Repr.pp

type closed_error = [

| `Closed

]

val closed_error_t : [ `Closed ] Irmin.Type.t

val pp_closed_error : [ `Closed ] Repr.pp

type read_only_error = [

| `Ro_not_allowed

]

val read_only_error_t : [ `Ro_not_allowed ] Irmin.Type.t

val pp_read_only_error : [ `Ro_not_allowed ] Repr.pp

type error = [

| base_error

| closed_error

| read_only_error

]

exception Pack_error of base_error

exception Closed

exception RO_not_allowed

module type S = sig ... end

Error manager

module Base : S with type t = error