Module Irmin_pack_unix.File_manager

module type S = sig ... end

Abstraction that governs the lifetime of the various files that are part of a pack store (except the branch store).

module Make (Io : Io.S) (Index : sig ... end) (Errs : Io_errors.S with module Io = Io) : S with module Io = Io and module Index = Index and module Errs = Errs

Abstraction that governs the lifetime of the various files that are part of a pack store (except the branch store).